Performed a import optmizer. Also overriden Max Health.

+++ src/main/java/com/airesnor/wuxiacraft/entities/skills/FireThrowable.java
@@ -19,116 +19,116 @@ import javax.annotation.Nullable;
 
 public class FireThrowable extends EntityThrowable {
 
-    private EntityLivingBase owner;
+	private EntityLivingBase owner;
 
-    private float damage;
+	private float damage;
 
-    private int duration;
+	private int duration;
 
-    private int particles;
+	private int particles;
 
-    public FireThrowable(World worldIn) {
-        super(worldIn);
-    }
+	public FireThrowable(World worldIn) {
+		super(worldIn);
+	}
 
-    public FireThrowable(World worldIn, EntityPlayer owner, float damage) {
-        this(worldIn, owner, damage, 20, 3, 0.08f);
-    }
+	public FireThrowable(World worldIn, EntityPlayer owner, float damage) {
+		this(worldIn, owner, damage, 20, 3, 0.08f);
+	}
 
-    public FireThrowable(World worldIn, EntityPlayer owner, float damage, int duration, int particles, float radius) {
-        super(worldIn, owner.posX, owner.posY + owner.getEyeHeight() - 0.1, owner.posZ);
-        this.setSize(radius, radius);
-        this.setNoGravity(true);
-        this.owner = owner;
-        this.thrower = this.owner;
-        this.damage = damage;
-        this.duration = duration;
-        this.particles = particles;
-    }
+	public FireThrowable(World worldIn, EntityPlayer owner, float damage, int duration, int particles, float radius) {
+		super(worldIn, owner.posX, owner.posY + owner.getEyeHeight() - 0.1, owner.posZ);
+		this.setSize(radius, radius);
+		this.setNoGravity(true);
+		this.owner = owner;
+		this.thrower = this.owner;
+		this.damage = damage;
+		this.duration = duration;
+		this.particles = particles;
+	}
 
-    @Override
-    public void shoot(Entity entityThrower, float rotationPitchIn, float rotationYawIn, float pitchOffset, float velocity, float inaccuracy) {
-        float f = -MathHelper.sin(rotationYawIn * 0.017453292F) * MathHelper.cos(rotationPitchIn * 0.017453292F);
-        float f1 = -MathHelper.sin((rotationPitchIn + pitchOffset) * 0.017453292F);
-        float f2 = MathHelper.cos(rotationYawIn * 0.017453292F) * MathHelper.cos(rotationPitchIn * 0.017453292F);
-        this.shoot((double)f, (double)f1, (double)f2, velocity, inaccuracy);
-    }
+	@Override
+	public void shoot(Entity entityThrower, float rotationPitchIn, float rotationYawIn, float pitchOffset, float velocity, float inaccuracy) {
+		float f = -MathHelper.sin(rotationYawIn * 0.017453292F) * MathHelper.cos(rotationPitchIn * 0.017453292F);
+		float f1 = -MathHelper.sin((rotationPitchIn + pitchOffset) * 0.017453292F);
+		float f2 = MathHelper.cos(rotationYawIn * 0.017453292F) * MathHelper.cos(rotationPitchIn * 0.017453292F);
+		this.shoot(f, f1, f2, velocity, inaccuracy);
+	}
 
-    @Override
-    public void onEntityUpdate() {
-        super.onEntityUpdate();
+	@Override
+	public void onEntityUpdate() {
+		super.onEntityUpdate();
 
-        if(this.ticksExisted >= this.duration || this.inWater) {
-            this.setDead();
-            return;
-        }
+		if (this.ticksExisted >= this.duration || this.inWater) {
+			this.setDead();
+			return;
+		}
 
-        if(this.ticksExisted >= 2 && this.world instanceof WorldServer) {
-            WorldServer worldServer = (WorldServer) this.world;
-            worldServer.spawnParticle(EnumParticleTypes.FLAME, false, this.posX, this.posY, this.posZ, this.particles, this.width, this.height, this.width, 0.002d, 0);
+		if (this.ticksExisted >= 2 && this.world instanceof WorldServer) {
+			WorldServer worldServer = (WorldServer) this.world;
+			worldServer.spawnParticle(EnumParticleTypes.FLAME, false, this.posX, this.posY, this.posZ, this.particles, this.width, this.height, this.width, 0.002d, 0);
 
-            AxisAlignedBB expandedBoundingBox = this.getEntityBoundingBox().grow(1, 1, 1);
-            worldServer.getEntitiesInAABBexcluding(this.owner, expandedBoundingBox, input -> !this.equals(input)).forEach(this::setEntityOnFire);
-        }
+			AxisAlignedBB expandedBoundingBox = this.getEntityBoundingBox().grow(1, 1, 1);
+			worldServer.getEntitiesInAABBexcluding(this.owner, expandedBoundingBox, input -> !this.equals(input)).forEach(this::setEntityOnFire);
+		}
 
-    }
+	}
 
-    @Override
-    protected void onImpact(RayTraceResult result) {
-        if(!this.world.isRemote) {
-            if(result.typeOfHit == RayTraceResult.Type.ENTITY && !result.entityHit.equals(this.owner)) {
-                this.attackEntityOnDirectHit((EntityLivingBase) result.entityHit);
-                this.setDead();
-            } else if(result.typeOfHit == RayTraceResult.Type.BLOCK) {
-                this.setHitBlockOnFire(result);
-                if(this.canNotPassThroughHitBlock(result)) {
-                    if(this.world instanceof WorldServer) {
-                        WorldServer worldServer = (WorldServer) this.world;
-                        worldServer.spawnParticle(EnumParticleTypes.FLAME, false, this.posX, this.posY, this.posZ, this.particles*3, this.width, this.height, this.width, this.rand.nextGaussian() / 10, 0);
-                    }
-                    this.setDead();
-                }
-            }
-        }
-    }
+	@Override
+	protected void onImpact(RayTraceResult result) {
+		if (!this.world.isRemote) {
+			if (result.typeOfHit == RayTraceResult.Type.ENTITY && !result.entityHit.equals(this.owner)) {
+				this.attackEntityOnDirectHit((EntityLivingBase) result.entityHit);
+				this.setDead();
+			} else if (result.typeOfHit == RayTraceResult.Type.BLOCK) {
+				this.setHitBlockOnFire(result);
+				if (this.canNotPassThroughHitBlock(result)) {
+					if (this.world instanceof WorldServer) {
+						WorldServer worldServer = (WorldServer) this.world;
+						worldServer.spawnParticle(EnumParticleTypes.FLAME, false, this.posX, this.posY, this.posZ, this.particles * 3, this.width, this.height, this.width, this.rand.nextGaussian() / 10, 0);
+					}
+					this.setDead();
+				}
+			}
+		}
+	}
 
-    private void attackEntityOnDirectHit(EntityLivingBase hitEntity) {
-        hitEntity.setFire(5);
-        hitEntity.attackEntityFrom(DamageSource.IN_FIRE, this.damage);
-        if(this.owner != null) {
-            hitEntity.setLastAttackedEntity(this.owner);
-        }
-    }
+	private void attackEntityOnDirectHit(EntityLivingBase hitEntity) {
+		hitEntity.setFire(5);
+		hitEntity.attackEntityFrom(DamageSource.IN_FIRE, this.damage);
+		if (this.owner != null) {
+			hitEntity.setLastAttackedEntity(this.owner);
+		}
+	}
 
-    private boolean canNotPassThroughHitBlock(RayTraceResult result) {
-        BlockPos hitBlockPos = result.getBlockPos();
-        IBlockState hitState = this.world.getBlockState(hitBlockPos);
-        return hitState.getMaterial().blocksMovement();
-    }
+	private boolean canNotPassThroughHitBlock(RayTraceResult result) {
+		BlockPos hitBlockPos = result.getBlockPos();
+		IBlockState hitState = this.world.getBlockState(hitBlockPos);
+		return hitState.getMaterial().blocksMovement();
+	}
 
-    private void setHitBlockOnFire(RayTraceResult result) {
-        BlockPos positionHit = result.getBlockPos();
-        IBlockState fire = Blocks.FIRE.getDefaultState();
-        // set the block this is on on fire
-        if(fire.getBlock().canPlaceBlockOnSide(this.world, positionHit, result.sideHit) && this.world.getBlockState(result.getBlockPos()).getMaterial().getCanBurn() && (!this.world.getBlockState(positionHit.offset(result.sideHit)).getMaterial().isLiquid())) {
-            BlockPos blockPosForFire = positionHit.offset(result.sideHit);
-            this.world.setBlockState(blockPosForFire, fire, 11);
-        }
-    }
+	private void setHitBlockOnFire(RayTraceResult result) {
+		BlockPos positionHit = result.getBlockPos();
+		IBlockState fire = Blocks.FIRE.getDefaultState();
+		// set the block this is on on fire
+		if (fire.getBlock().canPlaceBlockOnSide(this.world, positionHit, result.sideHit) && this.world.getBlockState(result.getBlockPos()).getMaterial().getCanBurn() && (!this.world.getBlockState(positionHit.offset(result.sideHit)).getMaterial().isLiquid())) {
+			BlockPos blockPosForFire = positionHit.offset(result.sideHit);
+			this.world.setBlockState(blockPosForFire, fire, 11);
+		}
+	}
 
-    private void setEntityOnFire(@Nullable Entity target) {
-        if(!target.isImmuneToFire() && !target.equals(this.owner)) {
-            target.setFire(5);
-        }
-    }
+	private void setEntityOnFire(@Nullable Entity target) {
+		if (!target.isImmuneToFire() && !target.equals(this.owner)) {
+			target.setFire(5);
+		}
+	}
 
-    @Override
-    public boolean isInvisible() {
-        return true;
-    }
+	@Override
+	public boolean isInvisible() {
+		return true;
+	}
 
-    @Override
-    public boolean shouldRenderInPass(int pass) {
-        return false;
-    }
+	@Override
+	public boolean shouldRenderInPass(int pass) {
+		return false;
+	}
 }
